requirements.txt
amqp==5.2.0
ansicon==1.89.0
arrow==1.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==24.3.0
beautifulsoup4==4.12.3
billiard==4.2.0
blessed==1.20.0
bs4==0.0.2
celery==5.4.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.5
Django==4.2.26
django-auth-adfs==1.14.0
django-cachalot==2.6.1
django-cors-headers==4.2.0
django-jsoneditor==0.2.4
django-picklefield==3.2
django-q==1.3.9
django-redis==5.4.0
django-unfold==0.46.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et_xmlfile==2.0.0
gunicorn==23.0.0
hiredis==2.2.3
idna==3.6
isodate==0.7.2
jinxed==1.3.0
kombu==5.3.7
lxml==5.3.0
mysqlclient==2.2.0
numpy==2.2.6
openpyxl==3.1.5
packaging==24.0
pandas==2.3.1
Pillow==10.3.0
platformdirs==4.3.6
prompt-toolkit==3.0.43
psutil==5.9.8
pycparser==2.22
PyJWT==2.8.0
pymemcache==4.0.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.4
redis==3.5.3
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
types-python-dateutil==2.9.0.20240821
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
zeep==4.3.1
